This version seems to work

+++ nengo_spinnaker/simulator.py
@@ -5,7 +5,8 @@ from nengo.cache import get_default_decoder_cache
 import numpy as np
 from rig.machine_control import MachineController
 from rig.machine_control.consts import AppState
-from rig.machine import Cores
+from rig.place_and_route import Cores
+import rig.place_and_route
 import six
 import time
 
@@ -13,8 +14,8 @@ from .builder import Model
 from .node_io import Ethernet
 from .rc import rc
 from .utils.config import getconfig
-from .utils.machine_control import test_and_boot
-from .utils import model as model_optimisations
+from .utils.model import (get_force_removal_passnodes,
+                          optimise_out_passthrough_nodes)
 
 logger = logging.getLogger(__name__)
 
@@ -49,7 +50,7 @@ class Simulator(object):
     def _remove_simulator(cls, simulator):
         cls._open_simulators.remove(simulator)
 
-    def __init__(self, network, dt=0.001, period=10.0):
+    def __init__(self, network, dt=0.001, period=10.0, timescale=1.0):
         """Create a new Simulator with the given network.
 
         Parameters
@@ -57,6 +58,9 @@ class Simulator(object):
         period : float or None
             Duration of one period of the simulator. This determines how much
             memory will be allocated to store precomputed and probed data.
+        timescale : float
+            Scaling factor to apply to the simulation, e.g., a value of `0.5`
+            will cause the simulation to run at half real-time.
         """
         # Add this simulator to the set of open simulators
         Simulator._add_simulator(self)
@@ -67,7 +71,7 @@ class Simulator(object):
         machine_height = rc.getint("spinnaker_machine", "height")
 
         self.controller = MachineController(hostname)
-        test_and_boot(self.controller, hostname, machine_width, machine_height)
+        self.controller.boot(machine_width, machine_height)
 
         # Create the IO controller
         io_cls = getconfig(network.config, Simulator, "node_io", Ethernet)
@@ -75,36 +79,41 @@ class Simulator(object):
                               dict())
         self.io_controller = io_cls(**io_kwargs)
 
+        # Calculate the machine timestep, this is measured in microseconds
+        # (hence the 1e6 scaling factor).
+        self.timescale = timescale
+        machine_timestep = int((dt / timescale) * 1e6)
+
         # Determine the maximum run-time
         self.max_steps = None if period is None else int(period / dt)
 
         self.steps = 0  # Steps simulated
 
-        # Create the IO controller. Function of time nodes are only enabled if
-        # the simulator period is not None.
-        io_cls = getconfig(network.config, Simulator, "node_io", Ethernet)
-        io_kwargs = getconfig(network.config, Simulator, "node_io_kwargs",
-                              dict())
-        self.io_controller = io_cls(
-            function_of_time_nodes=self.max_steps is not None, **io_kwargs
-        )
+        # If the simulator is in "run indefinite" mode (i.e., max_steps=None)
+        # then we modify the builders to ignore function of time Nodes and
+        # probes.
+        builder_kwargs = self.io_controller.builder_kwargs
+        if self.max_steps is None:
+            raise NotImplementedError
 
-        # Create a model from the network, using the IO controller. Probes are
-        # only built if the simulator period is not None.
+        # Create a model from the network, using the IO controller
         logger.debug("Building model")
         start_build = time.time()
-        self.model = Model(dt, decoder_cache=get_default_decoder_cache())
-        self.model.build(network, build_probes=self.max_steps is not None,
-                         **self.io_controller.builder_kwargs)
-        model_optimisations.remove_childless_filters(self.model)
+        self.model = Model(dt=dt, machine_timestep=machine_timestep,
+                           decoder_cache=get_default_decoder_cache())
+        self.model.build(network, **builder_kwargs)
+
+        forced_removals = get_force_removal_passnodes(network)
+        optimise_out_passthrough_nodes(self.model,
+                                       self.io_controller.passthrough_nodes,
+                                       network.config, forced_removals)
+
         logger.info("Build took {:.3f} seconds".format(time.time() -
                                                        start_build))
 
         self.model.decoder_cache.shrink()
         self.dt = self.model.dt
         self._closed = False  # Whether the simulator has been closed or not
-        self._running = False
-        self._halt = False
 
         self.host_sim = self._create_host_sim()
 
@@ -119,13 +128,19 @@ class Simulator(object):
         start = time.time()
         self.netlist = self.model.make_netlist(self.max_steps or 0)
 
-        # Get a machine object to place & route against
+        # Get a system-info object to place & route against
         logger.info("Getting SpiNNaker machine specification")
-        machine = self.controller.get_machine()
+        system_info = self.controller.get_system_info()
 
         # Place & Route
         logger.info("Placing and routing")
-        self.netlist.place_and_route(machine)
+        self.netlist.place_and_route(
+            system_info,
+            place=getconfig(network.config, Simulator,
+                            'placer', rig.place_and_route.place),
+            place_kwargs=getconfig(network.config, Simulator,
+                                   'placer_kwargs', {}),
+        )
 
         logger.info("{} cores in use".format(len(self.netlist.placements)))
         chips = set(six.itervalues(self.netlist.placements))
@@ -137,7 +152,7 @@ class Simulator(object):
 
         # Load the application
         logger.info("Loading application")
-        self.netlist.load_application(self.controller)
+        self.netlist.load_application(self.controller, system_info)
 
         # Check if any cores are in bad states
         if self.controller.count_cores_in_state(["exit", "dead", "watchdog",
@@ -159,7 +174,7 @@ class Simulator(object):
         for x in range(machine_width):
             for y in range(machine_height):
                 with self.controller(x=x, y=y):
-                    if (x, y) in machine:
+                    if (x, y) in system_info:
                         data = self.controller.read(0xf1000000, 4)
                         self.controller.write(0xf1000000, data[:-1] + b'\x10')
 
@@ -177,23 +192,16 @@ class Simulator(object):
 
     def run(self, time_in_seconds):
         """Simulate for the given length of time."""
-        if time_in_seconds is not None:
-            # Determine how many steps to simulate for
-            steps = int(np.round(float(time_in_seconds) / self.dt))
-            self.run_steps(steps)
-        else:
-            self._run_steps(None)
+        # Determine how many steps to simulate for
+        steps = int(np.round(float(time_in_seconds) / self.dt))
+        self.run_steps(steps)
 
     def run_steps(self, steps):
         """Simulate a give number of steps."""
-        if steps is not None and self.max_steps is not None:
-            # Fixed time simulation in either fixed-period or indefinite mode
-            while steps > 0:
-                n_steps = min((steps, self.max_steps))
-                self._run_steps(n_steps)
-                steps -= n_steps
-        else:
-            self._run_steps(steps)
+        while steps > 0:
+            n_steps = min((steps, self.max_steps))
+            self._run_steps(n_steps)
+            steps -= n_steps
 
     def _run_steps(self, steps):
         """Simulate for the given number of steps."""
@@ -201,135 +209,96 @@ class Simulator(object):
             raise Exception("Simulator has been closed and can't be used to "
                             "run further simulations.")
 
-        if self.max_steps is not None:
-            # The simulator is in "fixed-period, fixed-duration" mode.
+        if steps is None:
+            if self.max_steps is not None:
+                raise Exception(
+                    "Cannot run indefinitely if a simulator period was "
+                    "specified. Create a new simulator with Simulator(model, "
+                    "period=None) to perform indefinite time simulations."
+                )
+        else:
             assert steps <= self.max_steps
 
-            # Prepare the simulation
-            self.netlist.before_simulation(self, steps)
+        # Prepare the simulation
+        self.netlist.before_simulation(self, steps)
 
-            # Wait for all cores to hit SYNC0
+        # Wait for all cores to hit SYNC0 (either by remaining it or entering
+        # it from init)
+        self._wait_for_transition(AppState.init, AppState.sync0,
+                                  len(self.netlist.vertices))
+        self.controller.send_signal("sync0")
+
+        # Get a new thread for the IO
+        io_thread = self.io_controller.spawn()
+
+        # Run the simulation
+        try:
+            # Prep
+            exp_time = steps * self.dt / self.timescale
+            io_thread.start()
+
+            # Wait for all cores to hit SYNC1
             self.controller.wait_for_cores_to_reach_state(
-                "sync0", len(self.netlist.vertices)
+                "sync1", len(self.netlist.vertices)
             )
-            self.controller.send_signal("sync0")
+            logger.info("Running simulation...")
+            self.controller.send_signal("sync1")
 
-            # Get a new thread for the IO
-            io_thread = self.io_controller.spawn()
+            # Execute the local model
+            host_steps = 0
+            start_time = time.time()
+            run_time = 0.0
+            local_timestep = self.dt / self.timescale
+            while run_time < exp_time:
+                # Run a step
+                self.host_sim.step()
+                run_time = time.time() - start_time
 
-            # Run the simulation
-            self._running = True
-            try:
-                # Prep
-                exp_time = steps * (self.model.machine_timestep / float(1e6))
-                io_thread.start()
-
-                # Wait for all cores to hit SYNC1
-                self.controller.wait_for_cores_to_reach_state(
-                    "sync1", len(self.netlist.vertices)
-                )
-                logger.info("Running simulation...")
-                self.controller.send_signal("sync1")
-
-                # Execute the local model
-                host_steps = 0
-                start_time = time.time()
-                run_time = 0.0
-                while run_time < exp_time:
-                    # Run a step
-                    self.host_sim.step()
-                    run_time = time.time() - start_time
-
-                    # If that step took less than timestep then spin
+                # If that step took less than timestep then spin
+                time.sleep(0.0001)
+                while run_time < host_steps * local_timestep:
                     time.sleep(0.0001)
-                    while run_time < host_steps * self.dt:
-                        time.sleep(0.0001)
-                        run_time = time.time() - start_time
-            finally:
-                # Stop the IO thread whatever occurs
-                io_thread.stop()
-
-            # Check if any cores are in bad states
-            if self.controller.count_cores_in_state(["dead", "watchdog",
-                                                     "runtime_exception"]):
-                for vertex in self.netlist.vertices:
-                    x, y = self.netlist.placements[vertex]
-                    p = self.netlist.allocations[vertex][Cores].start
-                    status = self.controller.get_processor_status(p, x, y)
-                    if status.cpu_state is not AppState.sync0:
-                        print("Core ({}, {}, {}) in state {!s}".format(
-                            x, y, p, status.cpu_state))
-                raise Exception("Unexpected core failures.")
-
-            # Retrieve simulation data
-            start = time.time()
-            logger.info("Retrieving simulation data")
-            self.netlist.after_simulation(self, steps)
-            logger.info("Retrieving data took {:3f} seconds".format(
-                time.time() - start
-            ))
-
-            # Increase the steps count
-            self.steps += steps
-        else:
-            # The simulator is in "indefinite duration" mode.
-            if steps is not None:
-                raise ValueError("Cannot run an indefinite duration simulator "
-                                 "for a fixed period of time.")
-
-            # Prepare the simulation
-            self.netlist.before_simulation(self, None)
-
-            # Get a new thread for the IO
-            io_thread = self.io_controller.spawn()
-
-            if not self._running:
-                # If not already running then start things up
-                self.controller.wait_for_cores_to_reach_state(
-                    "sync0", len(self.netlist.vertices)
-                )
-                self.controller.send_signal("sync0")
-                self._running = True
-            else:
-                # Otherwise continue a running simulation
-                self.controller.send_signal("cont")
-
-            # Allow the local simulator to run
-            self._halt = False
-
-            # Run the simulation
-            try:
-                # Prep
-                io_thread.start()
-
-                # Wait for all cores to hit SYNC1
-                self.controller.wait_for_cores_to_reach_state(
-                    "sync1", len(self.netlist.vertices)
-                )
-                logger.info("Running simulation...")
-                self.controller.send_signal("sync1")
-
-                # Execute the local model
-                host_steps = 0
-                start_time = time.time()
-                run_time = 0.0
-                while not self._halt:
-                    # Run a step
-                    self.host_sim.step()
                     run_time = time.time() - start_time
+        finally:
+            # Stop the IO thread whatever occurs
+            io_thread.stop()
 
-                    # If that step took less than timestep then spin
-                    time.sleep(0.0001)
-                    while run_time < host_steps * self.dt:
-                        time.sleep(0.0001)
-                        run_time = time.time() - start_time
-            finally:
-                # Stop the IO thread whatever occurs
-                io_thread.stop()
+        # Wait for cores to re-enter sync0
+        self._wait_for_transition(AppState.run, AppState.sync0,
+                                  len(self.netlist.vertices))
 
-    def stop(self):
-        """Stop a continuously running simulation."""
-        self._halt = True
+        # Retrieve simulation data
+        start = time.time()
+        logger.info("Retrieving simulation data")
+        self.netlist.after_simulation(self, steps)
+        logger.info("Retrieving data took {:3f} seconds".format(
+            time.time() - start
+        ))
+
+        # Increase the steps count
+        self.steps += steps
+
+    def _wait_for_transition(self, from_state, desired_to_state, num_verts):
+        while True:
+            # If no cores are still in from_state, stop
+            if self.controller.count_cores_in_state(from_state) == 0:
+                break
+
+            # Wait a bit
+            time.sleep(1.0)
+
+        # Check if any cores haven't exited cleanly
+        if self.controller.count_cores_in_state(desired_to_state) != num_verts:
+            # Loop through all placed vertices
+            for vertex, (x, y) in six.iteritems(self.netlist.placements):
+                p = self.netlist.allocations[vertex][Cores].start
+                status = self.controller.get_processor_status(p, x, y)
+                if status.cpu_state is not desired_to_state:
+                    print("Core ({}, {}, {}) in state {!s}".format(
+                        x, y, p, status.cpu_state))
+                    print(self.controller.get_iobuf(p, x, y))
+            raise Exception("Unexpected core failures before reaching %s "
+                            "state." % desired_to_state)
 
     def _create_host_sim(self):
         # change node_functions to reflect time
@@ -345,13 +314,17 @@ class Simulator(object):
                         now = time.time()
                         if node_info['start'] is None:
                             node_info['start'] = now
-                        return f(now - node_info['start'])
+
+                        t = (now - node_info['start']) * self.timescale
+                        return f(t)
                 else:
                     def func(t, x, f=old_func):
                         now = time.time()
                         if node_info['start'] is None:
                             node_info['start'] = now
-                        return f(now - node_info['start'], x)
+
+                        t = (now - node_info['start']) * self.timescale
+                        return f(t, x)
                 node.output = func
                 node_functions[node] = old_func
 
@@ -366,15 +339,14 @@ class Simulator(object):
 
     def close(self):
         """Clean the SpiNNaker board and prevent further simulation."""
-        # Stop the application
-        self._closed = True
-        if not self._halt:
-            self.stop()
-        self.io_controller.close()
-        self.controller.send_signal("stop")
+        if not self._closed:
+            # Stop the application
+            self._closed = True
+            self.io_controller.close()
+            self.controller.send_signal("stop")
 
-        # Remove this simulator from the list of open simulators
-        Simulator._remove_simulator(self)
+            # Remove this simulator from the list of open simulators
+            Simulator._remove_simulator(self)
 
     def trange(self, dt=None):
         return np.arange(1, self.steps + 1) * (self.dt or dt)
